fops_bot/requirements.txt
discord.py
discordhealthcheck
requests
lxml
pytz
python_weather
pyyaml
psycopg2-binary
pillow
opencv-python
sqlalchemy
alembic
colorlog
faapi
pint
influxdb-client